pyproject.toml
[project]
name = "PersonalScripts"
description = "Add your description here"
version = "0.1.0"
readme = "README.md"
authors = [
    { name = "David Brownell", email = "github@DavidBrownell.com" }
]
requires-python = ">= 3.13"
dependencies = [
    "attrs>=25.4.0",
    "dbrownell-common>=0.14.7",
    "pyyaml>=6.0.3",
    "requests>=2.32.3",
    "requests-oauth>=0.4.1",
    "requests-oauthlib>=2.0.0",
    "typer>=0.21.0",
]

[dependency-groups]
dev = [
    "dbrownell-commitemojis>=0.3.0",
    "pre-commit>=4.2.0",
    "ruff>=0.11.9",
]

[tool.pytest.ini_options]
addopts = "--verbose -vv --capture=no --cov=PersonalScripts --cov-fail-under=95.0"

[tool.ruff]
line-length = 110

[tool.ruff.lint]
exclude = ["tests/**"]

select = ["ALL"]

ignore = [
    "ANN002", # Missing type annotation for `*args`
    "ANN003", # Missing type annotation for `**kwargs`
    "BLE001", # Do not catch blind exception: `Exception`
    "COM812", # Trailing comma missing
    "D100", # Missing docstring in public module
    "D105", # Missing docstring in magic method
    "D107", # Missing docstring in `__init__` method
    "D202", # No blank lines allowed after function docstring
    "E501", # Line too long
    "FIX002", # Line contains TODO, consider resolving the issue
    "I001", # Import block is un-sorted or un-formatted
    "N802", # Function name `xxx` should be lowercase
    "N999", # Invalid module name
    "RSE102", # Unnecessary parentheses on raise exception
    "S101", # Use of assert detected
    "TC006", # Add quotes to type expression in `typing.cast()`
    "TD002", # Missing author in TODO
    "TD003", # Missing issue link for this TODO
    "TRY002", # Create your own exception
    "TRY300", # Consider moving this statement to an `else` block
    "UP032", # Use f-string instead of `format` call
]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.pylint]
max-args = 10
max-branches = 20
max-returns = 20
